feat (JobDB): refactor submitJob
